requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==6.0.2
psycopg2-binary==2.9.11
reportlab==4.0.7
python-dotenv==1.0.0
flask-swagger-ui==4.11.1
gunicorn==21.2.0
